requirements.txt
haystack-ai
python-dotenv
pdfminer.six
streamlit